+++ exclude_test.go
@@ -348,8 +348,7 @@ func block(line, col, covered, uncovered int) prettycov.Block {
 func withBlocks(name string, blocks ...prettycov.Block) prettycov.FileCoverage {
 	item := prettycov.FileCoverage{File: name, Blocks: blocks}
 	for _, b := range blocks {
-		item.Coverage.Covered += b.Coverage.Covered
-		item.Coverage.Uncovered += b.Coverage.Uncovered
+		item.Coverage.Add(b.Coverage)
 	}
 
 	return item